NavBar Requested changes
Made a separate file for NavBar and called it into DashBoard
